Feat: Adding more business rules to facade service

+++ FacadePattern.API/DataTransferObjects/Coupon/CouponResponse.cs
@@ -4,5 +4,5 @@ public sealed class CouponResponse
 {
     public required int Id { get; set; }
     public required string Name { get; set; }
-    public required double DiscountPorcentage { get; set; }
+    public required decimal DiscountPorcentage { get; set; }
 }
