+++ FacadePattern.API/DataTransferObjects/Coupon/CouponSave.cs
@@ -1,4 +1,4 @@
 ﻿namespace FacadePattern.API.DataTransferObjects.Coupon;
 
 public sealed record CouponSave(string Name,
-                               double DiscountPorcentage);
+                                decimal DiscountPorcentage);
