+++ FacadePattern.API/Entities/Coupon.cs
@@ -4,5 +4,5 @@ public sealed class Coupon
 {
     public int Id { get; set; }
     public required string Name { get; set; }    
-    public required double DiscountPorcentage { get; set; }
+    public required decimal DiscountPorcentage { get; set; }
 }
